requirements.txt
# Core data processing
kagglehub[pandas-datasets]>=0.2.0
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.10.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optimization
pulp>=3.3.0

# Web Interface
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

